ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic-settings==2.4.0
python-multipart==0.0.9
pytest==8.3.2
httpx==0.27.0
langchain==0.2.11
langchain-openai==0.1.11
openai==1.35.9
pdfplumber==0.11.2
python-docx==0.8.11
